image_converter_mcp_server/requirements.txt
mcp>=1.0.0
fastmcp>=0.1.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
